shared_mem: drive tests via hist_build; retire hist_demo + shared_mem_build
