+++ tests/examples/test_hist_csim.py
@@ -13,7 +13,7 @@ from pathlib import Path
 
 import pytest
 
-from examples.shared_mem.shared_mem_build import CSIM_CASES, generate_vitis_sources
+from examples.shared_mem.hist_build import CSIM_CASES, generate_vitis_sources
 from pysilicon.toolchain import toolchain
 
 SHARED_DIR = Path(__file__).resolve().parents[2] / "examples" / "shared_mem"
